// src/lib/api/model/gamemode.ts
/**
 * RomgleWebApi
 * No description provided (generated by Openapi Generator https://github.com/openapitools/openapi-generator)
 *
 * The version of the OpenAPI document: 1.0
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type Gamemode = 'Normal' | 'Daily';

export const Gamemode = {
    Normal: 'Normal' as Gamemode,
    Daily: 'Daily' as Gamemode
};